pyproject.toml
[tool.poetry]
name = "webcntr"
version = "0.1.0"
description = ""
authors = ["Matti Luukkainen <mluukkai@iki.fi>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
flask = "^3.0.3"
robotframework = "^7.1.1"
robotframework-seleniumlibrary = "^6.6.1"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
getenv = "^0.2.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.3"
psycopg2 = "^2.9.10"
sqlalchemy = "^2.0.36"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pylint = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
